Issue #24562 move inheritance and discriminator entities to Data JPA tests
