+++ dev/io.openliberty.data.internal_fat_jpa/test-applications/DataJPATestApp/src/test/jakarta/data/jpa/web/WorkAddress.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
+ * Copyright (c) 2022,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,15 +10,19 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package test.jakarta.data.web;
+package test.jakarta.data.jpa.web;
 
-import jakarta.data.repository.DataRepository;
-import jakarta.data.repository.Repository;
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
 
 /**
- * The only purpose of this repository is to make the Jakarta Data provider aware
- * of the existence of the WorkAddress entity as a subtype of the ShippingAddress entity.
+ * Entity for testing inheritance with DiscriminatorValue.
  */
-@Repository
-public interface WorkAddresses extends DataRepository<WorkAddress, Long> {
+@DiscriminatorValue("Work")
+@Entity
+public class WorkAddress extends ShippingAddress {
+
+    public int floorNumber;
+
+    public String office;
 }
